requirements.txt
fitparse
influxdb
watchdog
python-dotenv
loguru
ipcalc
rich